refactor: migrate to PostgreSQL-only architecture
Complete migration from dual-database approach to PostgreSQL-only for all environments (dev, test, prod).

## Database Changes
- Remove SQLite support from SQLx configuration (Cargo.toml)
- Update all migrations to use PostgreSQL native types (TIMESTAMP, BOOLEAN, BIGINT, DOUBLE PRECISION)
- Use PostgreSQL placeholders ($1, $2, ...) consistently
- Update seed data with PostgreSQL-specific syntax (ON CONFLICT DO NOTHING)

## Timestamp Handling Fixes
- Fix all timestamp bindings: Utc::now().to_rfc3339() -> Utc::now().naive_utc()
- Remove unnecessary borrowing on NaiveDateTime (implements Copy trait)
- Files fixed: admin.rs, quote.rs, session.rs, ssr.rs, upload.rs, materials.rs

## API Improvements
- Add created_at field to CreateSessionResponse for accurate expiration calculation
- Update E2E tests to compare server-side timestamps (created_at vs expires_at)

## Configuration & Documentation
- Update .env.example with PostgreSQL credentials matching docker-compose
- Add PostgreSQL service to GitHub Actions CI workflow
- Remove SQLite references from README, Dockerfile, SECURITY_AUDIT
- Delete docker-compose.sqlite.yml
- Create ADR-005 documenting PostgreSQL-only decision

## Testing
- Fix E2E session expiration tests (e2e/api/sessions.spec.js, e2e/validation.spec.js)
- All 75 E2E tests passing
- All quality checks passing (fmt, clippy, lint, format)

+++ src/services/file_processor.rs
@@ -10,7 +10,7 @@ use std::path::Path;
 pub struct ProcessedModel {
     pub volume_cm3: f64,
     pub dimensions_mm: Dimensions,
-    pub triangle_count: i32,
+    pub triangle_count: i64,
     pub support_analysis: SupportAnalysis,
 }
 
@@ -108,7 +108,7 @@ pub fn process_stl_file(file_path: &str) -> Result<ProcessedModel, AppError> {
         })
         .collect();
 
-    let triangle_count = triangles.len() as i32;
+    let triangle_count = triangles.len() as i64;
     let volume_cm3 = calculate_volume(&triangles);
     let dimensions_mm = calculate_dimensions(&triangles);
     let support_analysis = analyze_supports(&triangles);
@@ -151,7 +151,7 @@ pub fn process_3mf_file(file_path: &str) -> Result<ProcessedModel, AppError> {
     // Parse the XML to extract mesh data
     let triangles = parse_3mf_mesh(&model_xml)?;
 
-    let triangle_count = triangles.len() as i32;
+    let triangle_count = triangles.len() as i64;
     let volume_cm3 = calculate_volume(&triangles);
     let dimensions_mm = calculate_dimensions(&triangles);
     let support_analysis = analyze_supports(&triangles);
